requirements.txt
nltk==3.7
pdfplumber==0.9.0
Pillow==9.4.0
Pillow==10.0.1
spacy==3.5.3
streamlit==1.26.0
